+++ msgraph/accesspackageresourcerequest_test.go
@@ -71,7 +71,7 @@ func testAccessPackageResourceRequestClient_Create(t *testing.T, c *test.Test, a
 		t.Fatal("AccessPackageResourceRequestClient.Create(): accessPackageResourceRequest was nil")
 	}
 	if accessPackageResourceRequest.ID == nil {
-		t.Fatal("AccessPackageResourceRequestClient.Create(): acccessPackageResourceRequest.ID was nil")
+		t.Fatal("AccessPackageResourceRequestClient.Create(): accessPackageResourceRequest.ID was nil")
 	}
 	return
 }
@@ -125,7 +125,7 @@ func testAccessPackageResourceRequestAP_Create(t *testing.T, c *test.Test, a msg
 		t.Fatal("AccessPackageClient.Create(): accessPackage was nil")
 	}
 	if accessPackage.ID == nil {
-		t.Fatal("AccessPackageClient.Create(): acccessPackage.ID was nil")
+		t.Fatal("AccessPackageClient.Create(): accessPackage.ID was nil")
 	}
 	return
 }