+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 0.39.0 (Unreleased)
+## 0.40.0 (Unreleased)
+
+## 0.39.0 (January 7, 2022)
+
+- Support for [Federated Identity Credentials](https://docs.microsoft.com/en-us/graph/api/resources/federatedidentitycredential?view=graph-rest-beta) (beta-only) ([#134](https://github.com/manicminer/hamilton/pull/134))
+- Bug fix: corrected the `DisplayName` struct tag for the `GroupAssignedLabel` model ([#135](https://github.com/manicminer/hamilton/pull/135))
+- Bug fix: fixed a typo in the constant `AccessPackageResourceRequestTypeAdminRemove` (was `AccessPackageResourceRequestTypeAdmminRemove`) ([#135](https://github.com/manicminer/hamilton/pull/135))
 
 ## 0.38.0 (December 8, 2021)